backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
networkx==3.4.2
numpy==2.1.3
scipy==1.14.1
httpx==0.28.0
